requirements.txt
python-binance
pandas
numpy
matplotlib
